+++ packages/store/experimental/src/index.ts
@@ -1,0 +1,1 @@
+export { withExperimentalNgxsPendingTasks } from './pending-tasks';
