+++ tsconfig.base.json
@@ -29,6 +29,7 @@
       "@ngxs/storage-plugin": ["packages/storage-plugin/index.ts"],
       "@ngxs/storage-plugin/internals": ["packages/storage-plugin/internals/src/index.ts"],
       "@ngxs/store": ["packages/store/index.ts"],
+      "@ngxs/store/experimental": ["packages/store/experimental/src/index.ts"],
       "@ngxs/store/internals": ["packages/store/internals/src/index.ts"],
       "@ngxs/store/internals/testing": ["packages/store/internals/testing/src/index.ts"],
       "@ngxs/store/operators": ["packages/store/operators/src/index.ts"],
